% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{return_isotope_df}
\alias{return_isotope_df}
\title{Isotope data frame}
\usage{
return_isotope_df(
  peaks,
  mz_col,
  rt_col,
  isotope_differences,
  mz_threshold,
  rt_threshold
)
}
\description{
Isotope data frame
}
