% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{return_isotope_df}
\alias{return_isotope_df}
\title{Isotope data frame}
\usage{
return_isotope_df(
  peaks,
  mz_col,
  rt_col,
  isotope_differences,
  mz_threshold,
  rt_threshold
)
}
\arguments{
\item{peaks}{dataframe with mz and rt info}

\item{mz_col}{mz column anme}

\item{rt_col}{rt column name}

\item{isotope_differences}{vector of isotope differences (mz diff)}

\item{mz_threshold}{Acceptable level of deviation from expected value}

\item{rt_threshold}{Acceptable level of deviation from expected value}
}
\value{
A dataframe with the isotopes
}
\description{
Isotope data frame
}
